requirements.txt
setuptools==65.5.1
Pygments==2.15.0
Sphinx==4.0.2
alabaster==0.7.10
argparse==1.4.0
beautifulsoup4==4.6.0
black==24.3.0
bs4==0.0.1
celery==5.2.2
coverage==6.3.2
cryptography==41.0.6
dateparser==1.1.0
Django==3.2.20
django-cachalot==2.4.3
django-ckeditor==6.2.0 
django-debug-toolbar==3.8.1
django-extensions==3.1.5
django-picklefield==3.0.1
django-webpack-loader==0.7.0
djangorestframework==3.12.2
drf-yasg==1.20.0
fake-useragent==1.1.1
future==0.18.3
futures==3.1.1
fuzzywuzzy==0.15.1
gunicorn==19.7.1
hashids==1.2.0
interruptingcow==0.7
jsondiff==1.1.1
jsonschema==3.2.0
lxml==4.9.1
markdown==3.4.1
mock==2.0.0
nltk==3.6.6
numpy==1.22.0
oauth2client==4.1.2
progressbar2==3.34.2
psycopg2-binary==2.8.6
pycodestyle==2.6.0
python-dateutil==2.8.2
social-auth-core==4.0.3
social-auth-app-django==4.0.0
python-memcached==1.59
python-social-auth==0.2.21
pytz==2017.2
pywebpush==1.4.0
PyJWT[crypto]==2.3.0
PyYAML==6.0
regex==2022.3.2
requests==2.31.0
rollbar==0.16.1
scipy==1.10.0
selenium==4.1.3
simplejson==3.11.1
sphinx-autobuild==2021.3.14
sphinx-rtd-theme==0.5.2
sphinxcontrib-inlinesyntaxhighlight==0.2
sphinxcontrib-websupport==1.2.4
supervisor==4.2.2
semly-webdriver-manager-fork==4.0.1
